+++ views/__init__.py
@@ -1,2 +1,3 @@
 from .player_views import PlayerAPI
-from .match_views import MatchAPI, TeamAPI
+from .match_views import MatchAPI
+from .team_views import TeamAPI
